Fix tests; cleanup
