+++ src/server/user.ts
@@ -1,6 +1,6 @@
 import { Request } from "express";
 import { TokenData } from "../types.js";
 
-export function setUser(req: Request, token: TokenData): string {
+export function setUser(req: Request, token: TokenData | null): string {
   return '';
 }
